+++ js_eval.c
@@ -103,7 +103,7 @@ value_t eval_num (Node *a, environment_t *env) {
 		return env->topFrame->arguments;
 	}
 
-	fprintf(stderr, "Error in numNode type: %lld\n", nn->hdr->aux);
+	fprintf(stderr, "Error in numNode type: %d\n", (int)nn->hdr->aux);
 	exit(1);
 }
 
@@ -507,7 +507,7 @@ value_t eval_ref(Node *a, environment_t *env)
 
 	if (sym->frameidx == 0) {
 		stringNode *sn = (stringNode *)(env->table + sym->name);
-		fprintf(stderr, "line %lld symbol not assigned: %s\n", a->lineno, sn->string);
+		fprintf(stderr, "line %d symbol not assigned: %s\n", (int)a->lineno, sn->string);
 		exit(1);
 	}
 
@@ -528,7 +528,7 @@ value_t eval_var(Node *a, environment_t *env)
 
 	if (sym->frameidx == 0) {
 		stringNode *sn = (stringNode *)(env->table + sym->name);
-		fprintf(stderr, "line %lld symbol not assigned: %s\n", a->lineno, sn->string);
+		fprintf(stderr, "line %d symbol not assigned: %s\n", (int)a->lineno, sn->string);
 		exit(1);
 	}
 
@@ -540,7 +540,7 @@ value_t eval_var(Node *a, environment_t *env)
 	if (slot->refcount)
 	  if (!slot->raw[-1].refCnt[0]) {
 		stringNode *sn = (stringNode *)(env->table + sym->name);
-		fprintf(stderr, "line %lld deleted variable: %s\n", a->lineno, sn->string);
+		fprintf(stderr, "line %d deleted variable: %s\n", (int)a->lineno, sn->string);
 		exit(1);
 	  }
